refactor(solver): split solver.py into 4 modules via mixins (P4)
Extract 1166 lines from solver.py (3338 -> 2171 lines) into three
mixin modules:
- external_regions.py (136 lines): ExternalRegionsMixin
- field_calculation.py (275 lines): FieldCalculationMixin
- layer_processing.py (673 lines): LayerProcessingMixin

FourierBaseSolver inherits from all three mixins, preserving the
public API and all import paths (from torchrdit.solver import ...).
577 tests pass, lint clean.
