+++ src/torchrdit/solver.py
@@ -114,17 +114,16 @@ Keywords:
     gradient-based optimization, PyTorch, computational electromagnetics
 """
 
-import os
-import time
 from typing import TYPE_CHECKING, Any, Callable, Dict, List, Optional, Tuple, Union
 
 import numpy as np
 import torch
-from torch.linalg import inv as tinv
-from torch.linalg import solve as tsolve
 
 from .algorithm import RCWAAlgorithm, RDITAlgorithm, SolverAlgorithm
-from .cell import Cell3D, CellType
+from .cell import Cell3D
+from .external_regions import ExternalRegionsMixin
+from .field_calculation import FieldCalculationMixin
+from .layer_processing import LayerProcessingMixin
 from .constants import Algorithm, Precision
 from .materials import MaterialClass
 from .numerics import safe_kz_reciprocal, softplus_floor, softplus_magnitude_floor
@@ -746,7 +745,7 @@ def softplus_clamp_min(x, min_val=0.0, beta=100):
     return softplus_floor(x, min_val, beta=beta)
 
 
-class FourierBaseSolver(Cell3D, SolverSubjectMixin):
+class FourierBaseSolver(Cell3D, SolverSubjectMixin, LayerProcessingMixin, FieldCalculationMixin, ExternalRegionsMixin):
     """Base class for Fourier-based electromagnetic solvers.
 
     This class serves as the foundation for Fourier-based electromagnetic solvers,
@@ -1351,282 +1350,6 @@ class FourierBaseSolver(Cell3D, SolverSubjectMixin):
                 if diff > tolerance:
                     print(f"[WARNING] {name} source {i} differs by more than {tolerance}")
 
-    def _process_layer(self, n_layer, matrices):
-        """Unified layer processing for both single and batched sources.
-
-        This method handles both homogeneous and non-homogeneous layers,
-        automatically detecting whether input is single source or batched.
-        It also respects the per-layer ``slice_count`` property by splitting
-        the layer thickness into identical slices and repeatedly composing the
-        resulting scattering matrices when ``slice_count > 1``.
-
-        Args:
-            n_layer: Layer index to process.
-            matrices: Dictionary of matrices from ``setup_common_matrices``.
-                     Can be either single source (3D) or batched (4D).
-
-        Returns:
-            dict: Scattering matrix for the layer with the same shape convention
-                  as the input (3D for single sources, 4D for batched sources).
-        """
-        # Matrices are always 4D: (n_sources, n_freqs, ...)
-        matrices_batched = matrices
-        n_sources = matrices_batched["mat_kx_diag"].shape[0]
-
-        if hasattr(self, "debug_tensorization") and self.debug_tensorization:
-            print(f"[DEBUG] _process_layer: layer {n_layer}, n_sources={n_sources}")
-            print(f"[DEBUG] mat_kx_diag shape: {matrices_batched['mat_kx_diag'].shape}")
-
-        layer = self.layer_manager.layers[n_layer]
-        slice_count = getattr(layer, "slice_count", 1)
-        if isinstance(layer.thickness, torch.Tensor):
-            layer_thickness_tensor = layer.thickness.to(self.device)
-        else:
-            layer_thickness_tensor = torch.tensor(
-                layer.thickness, dtype=self.tfloat, device=self.device
-            )
-        layer_thickness_complex = layer_thickness_tensor.to(self.tcomplex)
-        slice_thickness = layer_thickness_complex / slice_count  # Evenly split thickness across slices
-        smat_layer = None
-
-        # Handle non-homogeneous layers
-        if not layer.is_homogeneous:
-            # Extract needed matrices - now 4D: (n_sources, n_freqs, n_harmonics, n_harmonics)
-            mat_kx_diag = matrices_batched["mat_kx_diag"]
-            mat_ky_diag = matrices_batched["mat_ky_diag"]
-
-            if layer.is_dispersive:
-                toeplitz_er = layer.kermat
-                # Expand to match batch dimension if needed
-                if toeplitz_er.dim() == 3:
-                    toeplitz_er = toeplitz_er.unsqueeze(0).expand(n_sources, -1, -1, -1)
-            else:
-                toeplitz_er = self.expand_dims(layer.kermat)
-                if toeplitz_er.dim() == 3:
-                    toeplitz_er = toeplitz_er.unsqueeze(0).expand(n_sources, -1, -1, -1)
-
-            # assuming permeability always non-dispersive
-            toeplitz_ur = self.expand_dims(layer.kurmat)
-            if toeplitz_ur.dim() == 3:
-                toeplitz_ur = toeplitz_ur.unsqueeze(0).expand(n_sources, -1, -1, -1)
-
-            # Solve for all frequencies and sources (tsolve already supports 4D)
-            solve_ter_mky = tsolve(toeplitz_er, mat_ky_diag)
-            solve_ter_mkx = tsolve(toeplitz_er, mat_kx_diag)
-            solve_tur_mky = tsolve(toeplitz_ur, mat_ky_diag)
-            solve_tur_mkx = tsolve(toeplitz_ur, mat_kx_diag)
-
-            # Create P matrix (blockmat2x2 already supports 4D)
-            p_mat_i = blockmat2x2(
-                [
-                    [mat_kx_diag @ solve_ter_mky, toeplitz_ur - mat_kx_diag @ solve_ter_mkx],
-                    [mat_ky_diag @ solve_ter_mky - toeplitz_ur, -mat_ky_diag @ solve_ter_mkx],
-                ]
-            )
-
-            # Handle Fast Fourier Factorization (FFF)
-            if self.is_use_FFF:
-                if layer.mask_format is None:
-                    raise ValueError(f"Layer {n_layer} is missing mask data required for FFF.")
-
-                p_xx, p_yy, p_xy, p_yx = self._calculate_vector_field(
-                    mask=layer.mask_format.squeeze().to(self.tfloat).to(self.device),
-                    layer_index=n_layer,
-                )
-                reciprocal_toeplitz_er = self.layer_manager._gen_toeplitz2d(
-                    torch.reciprocal(layer.ermat.to(self.device).to(self.tcomplex)),
-                    nharmonic_1=self.harmonics[0],
-                    nharmonic_2=self.harmonics[1],
-                    method="FFT",
-                )
-
-                if reciprocal_toeplitz_er.dim() == 3:
-                    reciprocal_toeplitz_er = reciprocal_toeplitz_er.unsqueeze(0).expand(n_sources, -1, -1, -1)
-
-                delta_toeplitz_er = toeplitz_er - tinv(reciprocal_toeplitz_er)
-
-                if p_xy.dim() == 3:
-                    p_xy = p_xy.unsqueeze(0).expand(n_sources, -1, -1, -1)
-                    p_yx = p_yx.unsqueeze(0).expand(n_sources, -1, -1, -1)
-                    p_yy = p_yy.unsqueeze(0).expand(n_sources, -1, -1, -1)
-                    p_xx = p_xx.unsqueeze(0).expand(n_sources, -1, -1, -1)
-
-                # Liu 2012 eq.51 defines E2 in the [-Ey, Ex] field ordering,
-                # but torchrdit's Q matrix uses the [Ex, Ey] ordering.  The
-                # coordinate transformation swaps the diagonal P blocks:
-                #   Q[0:n, n:] uses p_xx (not p_yy)
-                #   Q[n:, 0:n] uses p_yy (not p_xx)
-                q_mat_i = blockmat2x2(
-                    [
-                        [
-                            mat_kx_diag @ solve_tur_mky - delta_toeplitz_er @ p_xy,
-                            toeplitz_er - mat_kx_diag @ solve_tur_mkx - delta_toeplitz_er @ p_xx,
-                        ],
-                        [
-                            mat_ky_diag @ solve_tur_mky - toeplitz_er + delta_toeplitz_er @ p_yy,
-                            delta_toeplitz_er @ p_yx - mat_ky_diag @ solve_tur_mkx,
-                        ],
-                    ]
-                )
-            else:
-                q_mat_i = blockmat2x2(
-                    [
-                        [mat_kx_diag @ solve_tur_mky, toeplitz_er - mat_kx_diag @ solve_tur_mkx],
-                        [mat_ky_diag @ solve_tur_mky - toeplitz_er, -mat_ky_diag @ solve_tur_mkx],
-                    ]
-                )
-
-            # Non-homogeneous layer — algorithms accept arbitrary leading batch dims
-            k_0_batched = self.k_0.unsqueeze(0).expand(n_sources, -1)  # (B, F)
-
-            smat_layer = self._solve_nonhomo_layer(
-                layer_thickness=slice_thickness,
-                p_mat_i=p_mat_i,                       # (B, F, M, N)
-                q_mat_i=q_mat_i,                       # (B, F, M, N)
-                mat_w0=matrices_batched["mat_w0"],     # (B, F, M, N)
-                mat_v0=matrices_batched["mat_v0"],     # (B, F, M, N)
-                harmonics=self.harmonics,
-                k_0=k_0_batched,                       # (B, F)
-            )
-
-        # Handle homogeneous layers
-        else:
-            # Debug print for tensor shapes
-            if hasattr(self, "debug_tensorization") and self.debug_tensorization:
-                print(f"[DEBUG] Processing homogeneous layer {n_layer}")
-                print(f"[DEBUG] mat_kx shape: {matrices_batched['mat_kx'].shape}")
-
-            if layer.is_dispersive:
-                # Get material properties for all frequencies
-                toeplitz_er = self._matlib[layer.material_name].er.detach().to(self.tcomplex).to(self.device)
-                toeplitz_ur = self._matlib[layer.material_name].ur.detach().to(self.tcomplex).to(self.device)
-
-                # Debug logging
-                if hasattr(self, "debug_tensorization") and self.debug_tensorization:
-                    print(
-                        f"[DEBUG] Dispersive layer - toeplitz_er shape: {toeplitz_er.shape}, dim: {toeplitz_er.dim()}"
-                    )
-                    print(
-                        f"[DEBUG] Dispersive layer - toeplitz_ur shape: {toeplitz_ur.shape}, dim: {toeplitz_ur.dim()}"
-                    )
-
-                # Handle scalar case (single wavelength)
-                if toeplitz_er.dim() == 0:
-                    toeplitz_er = toeplitz_er.unsqueeze(0)  # Make it 1D
-                    toeplitz_ur = toeplitz_ur.unsqueeze(0)  # Make it 1D
-
-                # Always expand to (n_sources, n_freqs) for uniform broadcasting
-                toeplitz_er = toeplitz_er.unsqueeze(0).expand(n_sources, -1)
-                toeplitz_ur = toeplitz_ur.unsqueeze(0).expand(n_sources, -1)
-
-                toep_ur_er = toeplitz_ur * toeplitz_er
-                conj_toep_ur_er = torch.conj(toeplitz_ur) * torch.conj(toeplitz_er)
-
-                # kz for the layer: (n_sources, n_freqs, n_harmonics²)
-                mat_kz_i = torch.conj(
-                    torch.sqrt(
-                        conj_toep_ur_er[:, :, None]
-                        - matrices_batched["mat_kx"] ** 2
-                        - matrices_batched["mat_ky"] ** 2
-                    )
-                    + 0 * 1j
-                ).to(self.tcomplex)
-
-                # Differentiable kz protection (see numerics.py)
-                inv_dmat_lam_i = safe_kz_reciprocal(mat_kz_i, dtype=self.tcomplex)
-
-                mat_v_i = (
-                    1
-                    / toeplitz_ur[:, :, None]
-                    * blockmat2x2(
-                        [
-                            [
-                                to_diag_util(matrices_batched["mat_kx_ky"] * inv_dmat_lam_i, self.harmonics),
-                                to_diag_util(
-                                    (toep_ur_er[:, :, None] - matrices_batched["mat_kx_kx"]) * inv_dmat_lam_i,
-                                    self.harmonics,
-                                ),
-                            ],
-                            [
-                                to_diag_util(
-                                    (matrices_batched["mat_ky_ky"] - toep_ur_er[:, :, None]) * inv_dmat_lam_i,
-                                    self.harmonics,
-                                ),
-                                -to_diag_util(matrices_batched["mat_ky_kx"] * inv_dmat_lam_i, self.harmonics),
-                            ],
-                        ]
-                    )
-                )
-            else:
-                # Get non-dispersive material properties
-                toeplitz_er = self._matlib[layer.material_name].er.detach().clone().to(self.device)
-                toeplitz_ur = self._matlib[layer.material_name].ur.detach().clone().to(self.device)
-
-                # Calculate common values
-                toep_ur_er = toeplitz_ur * toeplitz_er
-                conj_toep_ur_er = torch.conj(toeplitz_ur) * torch.conj(toeplitz_er)
-
-                mat_kz_i = torch.conj(
-                    torch.sqrt(conj_toep_ur_er - matrices_batched["mat_kx"] ** 2 - matrices_batched["mat_ky"] ** 2)
-                    + 0 * 1j
-                ).to(self.tcomplex)
-
-                # Differentiable kz protection (see numerics.py)
-                inv_dmat_lam_i = safe_kz_reciprocal(mat_kz_i, dtype=self.tcomplex)
-                mat_v_i = (
-                    1
-                    / toeplitz_ur
-                    * blockmat2x2(
-                        [
-                            [
-                                to_diag_util(matrices_batched["mat_kx_ky"] * inv_dmat_lam_i, self.harmonics),
-                                to_diag_util((toep_ur_er - matrices_batched["mat_kx_kx"]) * inv_dmat_lam_i, self.harmonics),
-                            ],
-                            [
-                                to_diag_util((matrices_batched["mat_ky_ky"] - toep_ur_er) * inv_dmat_lam_i, self.harmonics),
-                                -to_diag_util(matrices_batched["mat_ky_kx"] * inv_dmat_lam_i, self.harmonics),
-                            ],
-                        ]
-                    )
-                )
-
-            # Layer matrix: always (n_sources, n_freqs, n_harmonics²)
-            k_0_expanded = self.k_0.unsqueeze(0).unsqueeze(2)  # (1, n_freqs, 1)
-            mat_x_i = -1j * mat_kz_i * k_0_expanded * slice_thickness
-            mat_x_i_diag = torch.concat([mat_x_i, mat_x_i], dim=-1)
-
-            # Debug shapes before to_diag_util
-            if hasattr(self, "debug_tensorization") and self.debug_tensorization:
-                print(f"[DEBUG] mat_x_i shape: {mat_x_i.shape}")
-                print(f"[DEBUG] mat_x_i_diag shape: {mat_x_i_diag.shape}")
-                print(f"[DEBUG] harmonics: {self.harmonics}")
-
-            mat_x_i = to_diag_util(torch.exp(mat_x_i_diag), self.harmonics)
-
-            # Calculate Layer Scattering Matrix
-            atwi = matrices_batched["mat_w0"]
-            atvi = tsolve(mat_v_i, matrices_batched["mat_v0"])
-            mat_a_i = atwi + atvi
-            mat_b_i = atwi - atvi
-
-            solve_ai_xi = tsolve(mat_a_i, mat_x_i)
-
-            mat_xi_bi = mat_x_i @ mat_b_i
-
-            mat_d_i = mat_a_i - mat_xi_bi @ solve_ai_xi @ mat_b_i
-
-            s11 = tsolve(mat_d_i, mat_xi_bi @ solve_ai_xi @ mat_a_i - mat_b_i)
-            s12 = tsolve(mat_d_i, mat_x_i) @ (mat_a_i - mat_b_i @ tsolve(mat_a_i, mat_b_i))
-            smat_layer = SMatrix(S11=s11, S12=s12, S21=s12, S22=s11)
-
-        if slice_count > 1:
-            smat_combined = smat_layer
-            for _ in range(slice_count - 1):
-                smat_combined = redhstar(smat_combined, smat_layer, tcomplex=self.tcomplex)
-            smat_layer = smat_combined
-
-        return smat_layer
-
     def _solve_structure(self, sources=None, **kwargs):
         """Solve the electromagnetic problem for single or batched sources.
 
@@ -2102,896 +1825,6 @@ class FourierBaseSolver(Cell3D, SolverSubjectMixin):
                     # if no pattern assigned before solving, the material will be set as homogeneous
                     self.layer_manager.replace_layer_to_homogeneous(layer_index=n_layer)
                     print(f"Warning: Layer {n_layer} has no pattern assigned, and was changed to homogeneous")
-
-    def update_er_with_mask(
-        self, mask: torch.Tensor, layer_index: int, bg_material: str = "air", method: str = "FFT"
-    ) -> None:
-        """Update the permittivity distribution in a layer using a binary mask.
-
-        This method allows you to define the spatial distribution of materials in a layer
-        using a binary mask. The mask defines regions where the foreground material (mask=1)
-        and background material (mask=0) are located. This is particularly useful for
-        defining complex patterns, gratings, or arbitrary shapes within a layer.
-
-        This functionality is a key component for meta-optics design, enabling the creation of:
-        - Parameter-constrained and free-form meta-atoms
-        - Reconfigurable photonic structures
-        - Meta-lenses and beam deflectors
-        - Complex diffractive elements
-
-        When combined with automatic differentiation, this method enables gradient-based
-        optimization for the inverse design of meta-optics devices, as demonstrated in
-        "Eigendecomposition-free inverse design of meta-optics devices" (Huang et al., 2024).
-
-        If the specified layer is currently homogeneous, it will be automatically
-        converted to a grating (patterned) layer. The method computes the Fourier
-        transform of the pattern to obtain the permittivity distribution in k-space,
-        which is required for the RCWA/R-DIT calculations.
-
-        The method supports two approaches for computing the Toeplitz matrix:
-        - 'FFT': Uses Fast Fourier Transform, suitable for all cell types
-        - 'Analytical': Uses analytical expressions, only available for Cartesian cells
-
-        Args:
-            mask: Binary tensor representing the pattern mask, where:
-                 - 1 (or True) represents the foreground material (from the layer's material)
-                 - 0 (or False) represents the background material (specified by bg_material)
-                 The mask dimensions must match the real-space dimensions (grids) of the solver.
-
-                 For inverse design applications, this mask can be a differentiable tensor
-                 generated from a neural network or optimization process.
-
-            layer_index: Index of the layer to update. This should be a valid index in the
-                        layer stack (0 to nlayer-1).
-
-            bg_material: Name of the background material to use where mask=0. Default is 'air'.
-                        This must be a valid material name that exists in the solver's material list.
-
-            method: Method for computing the Toeplitz matrix:
-                   - 'FFT': Fast Fourier Transform method (works for all cell types)
-                   - 'Analytical': Analytical method (only works for Cartesian cells)
-                   Default is 'FFT'.
-
-        Returns:
-            None
-
-        Raises:
-            ValueError: If the mask dimensions don't match the solver's real-space dimensions.
-            KeyError: If the specified background material doesn't exist in the material list.
-
-        Examples:
-        ```python
-        # Basic usage: Create a binary mask for a simple grating pattern
-        mask = torch.zeros(512, 512)
-        mask[:, :256] = 1.0  # Half the domain has the foreground material
-
-        # Update layer 1 with the mask, using air as background
-        solver.update_er_with_mask(mask=mask, layer_index=1, bg_material='air')
-        ```
-
-        Inverse design usage:
-        ```python
-        import torch.nn as nn
-
-        # Define a simple network to generate a mask
-        class MaskGenerator(nn.Module):
-            def __init__(self, dim=512):
-                super().__init__()
-                self.layers = nn.Sequential(
-                    nn.Linear(2, 128),
-                    nn.ReLU(),
-                    nn.Linear(128, 1),
-                    nn.Sigmoid()  # Output in [0, 1] range
-                )
-                # Create normalized coordinate grid
-                x = torch.linspace(-1, 1, dim)
-                y = torch.linspace(-1, 1, dim)
-                self.X, self.Y = torch.meshgrid(x, y, indexing='ij')
-
-            def forward(self):
-                # Stack coordinates
-                coords = torch.stack([self.X.flatten(), self.Y.flatten()], dim=1)
-                # Generate mask values
-                mask_flat = self.layers(coords)
-                # Reshape to 2D
-                return mask_flat.reshape(self.X.shape)
-
-        # Create and use the mask generator
-        mask_gen = MaskGenerator(dim=512).to(device)
-        optimizer = torch.optim.Adam(mask_gen.parameters(), lr=0.001)
-
-        # Optimization loop
-        for i in range(100):
-            optimizer.zero_grad()
-
-            # Generate mask
-            mask = mask_gen()
-
-            # Update solver with mask
-            solver.update_er_with_mask(mask, layer_index=1)
-
-            # Solve and calculate loss
-            result = solver.solve(source) # SolverResults object
-            loss = -result.transmission[0]  # Maximize transmission
-
-            # Backpropagate and update
-            loss.backward()
-            optimizer.step()
-        ```
-
-        Note:
-            This method only updates the permittivity (epsilon) distribution. To update
-            the permeability (mu) distribution, you would need to modify the layer's
-            urmat property directly.
-
-        References:
-            - Huang et al., "Eigendecomposition-free inverse design of meta-optics devices,"
-              Opt. Express 32, 13986-13997 (2024)
-            - Huang et al., "Inverse Design of Photonic Structures Using Automatic Differentiable
-              Rigorous Diffraction Interface Theory," CLEO (2023)
-        """
-
-        ndim1, ndim2 = mask.size()
-        if (ndim1 != self.grids[0]) or (ndim2 != self.grids[1]):
-            raise ValueError("Mask dims don't match!")
-
-        if self.layer_manager.layers[layer_index].is_homogeneous:
-            self.layer_manager.replace_layer_to_grating(layer_index=layer_index)
-
-        er_bg = self._get_bg(layer_index=layer_index, param="er")
-
-        if self.layer_manager.layers[layer_index].is_dispersive is False:
-            mask_format = mask
-        else:
-            mask_format = mask.unsqueeze(-3)
-
-        self.layer_manager.layers[layer_index].mask_format = mask_format.to(self.tfloat)
-
-        # Store the background material in the layer for visualization
-        if isinstance(bg_material, str):
-            self.layer_manager.layers[layer_index].bg_material = bg_material
-        elif hasattr(bg_material, "name"):
-            self.layer_manager.layers[layer_index].bg_material = bg_material.name
-        else:
-            self.layer_manager.layers[layer_index].bg_material = str(bg_material)
-
-        if isinstance(bg_material, str) and bg_material.lower() == "air":
-            self.layer_manager.layers[layer_index].ermat = 1 + (er_bg - 1) * mask_format
-        elif isinstance(bg_material, MaterialClass) and bg_material.name.lower() == "air":
-            self.layer_manager.layers[layer_index].ermat = 1 + (er_bg - 1) * mask_format
-        else:
-            if self._matlib[bg_material].er.ndim == 0:
-                self.layer_manager.layers[layer_index].ermat = (
-                    self._matlib[bg_material].er * (1 - mask_format) + (er_bg - 1) * mask_format
-                )
-            else:
-                self.layer_manager.layers[layer_index].ermat = (
-                    self._matlib[bg_material].er[:, None, None] * (1 - mask_format) + (er_bg - 1) * mask_format
-                )
-
-        if method == "Analytical":
-            if self.cell_type != CellType.Cartesian:
-                print(f"method [{method}] does not support the cell type [{self.cell_type}], will use FFT instead.")
-                method = "FFT"
-        elif method != "Analytical" and method != "FFT":
-            method = "FFT"
-
-        self.layer_manager.gen_toeplitz_matrix(
-            layer_index=layer_index, n_harmonic1=self.harmonics[0], n_harmonic2=self.harmonics[1], param="er", method=method
-        )
-
-        # permeability always 1.0 for current version
-        # currently no support for magnetic materials
-        self.layer_manager.layers[layer_index].urmat = self._get_bg(layer_index=layer_index, param="ur")
-        self.layer_manager.gen_toeplitz_matrix(
-            layer_index=layer_index, n_harmonic1=self.harmonics[0], n_harmonic2=self.harmonics[1], param="ur", method=method
-        )
-
-
-    def _vector_generator_cache_key(
-        self,
-        *,
-        device: Union[str, torch.device],
-        dtype: torch.dtype,
-        fourier_loss_weight: float,
-        smoothness_loss_weight: float,
-        steps: int,
-    ) -> Tuple[
-        Tuple[int, int],
-        Tuple[float, ...],
-        Tuple[float, ...],
-        str,
-        torch.dtype,
-        float,
-        float,
-        int,
-    ]:
-        lattice_t1_vals = tuple(
-            float(x) for x in self.lattice_t1.detach().cpu().view(-1).tolist()
-        )
-        lattice_t2_vals = tuple(
-            float(x) for x in self.lattice_t2.detach().cpu().view(-1).tolist()
-        )
-        device_str = str(torch.device(device))
-        return (
-            tuple(int(k) for k in self.harmonics),
-            lattice_t1_vals,
-            lattice_t2_vals,
-            device_str,
-            dtype,
-            float(fourier_loss_weight),
-            float(smoothness_loss_weight),
-            int(steps),
-        )
-
-    def _get_tangent_field_generator(
-        self,
-        *,
-        device: Union[str, torch.device],
-        dtype: torch.dtype,
-        fourier_loss_weight: float,
-        smoothness_loss_weight: float,
-        steps: int,
-    ):
-        key = self._vector_generator_cache_key(
-            device=device,
-            dtype=dtype,
-            fourier_loss_weight=fourier_loss_weight,
-            smoothness_loss_weight=smoothness_loss_weight,
-            steps=steps,
-        )
-        generator = self._vector_generator_cache.get(key)
-        device_obj = torch.device(device)
-
-        if generator is None:
-            from . import vector as vector_module  # Local import to avoid circular deps
-
-            generator = vector_module.TangentFieldGenerator(
-                lattice_t1=self.lattice_t1.to(device=device_obj, dtype=dtype),
-                lattice_t2=self.lattice_t2.to(device=device_obj, dtype=dtype),
-                harmonics=tuple(int(k) for k in self.harmonics),
-                fourier_loss_weight=fourier_loss_weight,
-                smoothness_loss_weight=smoothness_loss_weight,
-                steps=steps,
-            )
-            self._vector_generator_cache[key] = generator
-        else:
-            generator._expansion_manager.adapt_to(device=device_obj, dtype=dtype)
-        return generator
-
-    def _calculate_vector_field(self, mask: torch.Tensor, *, layer_index: Optional[int] = None):
-        """Generate Fourier-factorization P-matrix components via tangent field."""
-        debug_enabled = getattr(self, "debug_fff", False)
-        if not debug_enabled:
-            env_flag = os.getenv("TORCHRDIT_DEBUG_FFF")
-            if env_flag:
-                debug_enabled = env_flag.strip().lower() not in {"0", "false", "off"}
-        start_time = time.perf_counter() if debug_enabled else None
-
-        squeezed_mask = mask.squeeze()
-        if squeezed_mask.dim() == 2:
-            batched_mask = squeezed_mask.unsqueeze(0)
-        elif squeezed_mask.dim() == 3:
-            batched_mask = squeezed_mask
-        else:
-            raise ValueError(f"Expected mask to squeeze to 2D or 3D, got shape {mask.shape}")
-
-        batched_mask = batched_mask.to(self.tfloat).to(self.device)
-
-        scheme = getattr(self, "fff_vector_scheme", getattr(self, "vector_field_scheme", "POL"))
-        fourier_loss_weight = getattr(self, "fff_fourier_weight", getattr(self, "vector_fourier_loss_weight", 1e-2))
-        smoothness_loss_weight = getattr(
-            self, "fff_smoothness_weight", getattr(self, "vector_smoothness_loss_weight", 1e-3)
-        )
-        steps = getattr(self, "fff_vector_steps", getattr(self, "vector_field_steps", 1))
-
-        generator = self._get_tangent_field_generator(
-            device=batched_mask.device,
-            dtype=batched_mask.dtype,
-            fourier_loss_weight=fourier_loss_weight,
-            smoothness_loss_weight=smoothness_loss_weight,
-            steps=steps,
-        )
-
-        tx_components: List[torch.Tensor] = []
-        ty_components: List[torch.Tensor] = []
-        for mask_sample in batched_mask:
-            tx_sample, ty_sample = generator.compute(
-                mask=mask_sample,
-                XO=self.XO,
-                YO=self.YO,
-                scheme=scheme,
-                fourier_loss_weight=fourier_loss_weight,
-                smoothness_loss_weight=smoothness_loss_weight,
-                steps=steps,
-            )
-            tx_components.append(tx_sample)
-            ty_components.append(ty_sample)
-
-        tx_stack = torch.stack(tx_components, dim=0)
-        ty_stack = torch.stack(ty_components, dim=0)
-
-        if tx_stack.shape[0] == 1:
-            tx = tx_stack[0]
-            ty = ty_stack[0]
-        else:
-            tx = tx_stack
-            ty = ty_stack
-
-        tx = tx.to(self.tcomplex)
-        ty = ty.to(self.tcomplex)
-        tx_conj = torch.conj(tx)
-        ty_conj = torch.conj(ty)
-
-        tx_mag_sq = (tx_conj * tx).real
-        ty_mag_sq = (ty_conj * ty).real
-        denom = torch.clamp(tx_mag_sq + ty_mag_sq, min=1e-12)
-        inv_denom = denom.reciprocal()
-        inv_denom_complex = inv_denom.to(dtype=self.tcomplex)
-
-        weight_xx = (tx_mag_sq * inv_denom).to(dtype=self.tcomplex)
-        weight_yy = (ty_mag_sq * inv_denom).to(dtype=self.tcomplex)
-        weight_xy = (tx * ty_conj) * inv_denom_complex
-        weight_yx = torch.conj(weight_xy)
-
-        toeplitz_xx = self.layer_manager._gen_toeplitz2d(
-            weight_xx, nharmonic_1=self.harmonics[0], nharmonic_2=self.harmonics[1], method="FFT"
-        )
-        toeplitz_yy = self.layer_manager._gen_toeplitz2d(
-            weight_yy, nharmonic_1=self.harmonics[0], nharmonic_2=self.harmonics[1], method="FFT"
-        )
-        toeplitz_xy = self.layer_manager._gen_toeplitz2d(
-            weight_xy, nharmonic_1=self.harmonics[0], nharmonic_2=self.harmonics[1], method="FFT"
-        )
-        toeplitz_yx = self.layer_manager._gen_toeplitz2d(
-            weight_yx, nharmonic_1=self.harmonics[0], nharmonic_2=self.harmonics[1], method="FFT"
-        )
-
-        if debug_enabled:
-            duration_ms = (time.perf_counter() - start_time) * 1000.0
-            n_harmonics_sq = self.harmonics[0] * self.harmonics[1]
-            bytes_per_complex = torch.tensor([], dtype=self.tcomplex).element_size()
-            approx_mem_mb = (4 * n_harmonics_sq * n_harmonics_sq * bytes_per_complex) / (1024.0 * 1024.0)
-            layer_info = f" layer {layer_index}" if layer_index is not None else ""
-            print(
-                f"[FFF DEBUG] calculate_vector_field{layer_info}: {duration_ms:.2f} ms, "
-                f"P-matrix approx {approx_mem_mb:.2f} MB"
-            )
-
-        return toeplitz_xx, toeplitz_yy, toeplitz_xy, toeplitz_yx
-
-    def get_vector_components(self, layer_index: int):
-        if self.layer_manager.layers[layer_index].mask_format is not None:
-            mask = self.layer_manager.layers[layer_index].mask_format.squeeze()
-            if mask.dim() != 2:
-                mask = mask.to(self.tfloat).to(self.device).squeeze()
-            else:
-                mask = mask.to(self.tfloat).to(self.device)
-
-            scheme = getattr(self, "fff_vector_scheme", getattr(self, "vector_field_scheme", "POL"))
-            fourier_loss_weight = getattr(
-                self, "fff_fourier_weight", getattr(self, "vector_fourier_loss_weight", 1e-2)
-            )
-            smoothness_loss_weight = getattr(
-                self, "fff_smoothness_weight", getattr(self, "vector_smoothness_loss_weight", 1e-3)
-            )
-            steps = getattr(self, "fff_vector_steps", getattr(self, "vector_field_steps", 1))
-
-            generator = self._get_tangent_field_generator(
-                device=mask.device,
-                dtype=mask.dtype,
-                fourier_loss_weight=fourier_loss_weight,
-                smoothness_loss_weight=smoothness_loss_weight,
-                steps=steps,
-            )
-
-            tx, ty = generator.compute(
-                mask=mask,
-                XO=self.XO,
-                YO=self.YO,
-                scheme=scheme,
-                fourier_loss_weight=fourier_loss_weight,
-                smoothness_loss_weight=smoothness_loss_weight,
-                steps=steps,
-            )
-
-            return tx, ty
-        else:
-            return None, None
-
-    def update_layer_thickness(self, layer_index: int, thickness: torch.Tensor):
-        """Update the thickness of a specific layer in the structure.
-
-        This method allows you to dynamically change the thickness of a layer
-        in the simulation structure. This is particularly useful for:
-        - Parametric sweeps over layer thicknesses
-        - Optimization of layer thicknesses for specific optical responses
-        - Dynamic adjustment of device structures during simulation
-
-        The thickness parameter can be a regular Tensor or one with gradients
-        enabled for automatic differentiation in optimization workflows.
-
-        Args:
-            layer_index (int): Index of the layer to modify. Must be a valid index
-                        in the range 0 to (number of layers - 1).
-            thickness (torch.Tensor): New thickness value for the layer.
-                        Must be a positive scalar Tensor in the solver's length units.
-                        Can have requires_grad=True for gradient-based optimization.
-
-        Examples:
-        ```python
-        import torch
-        from torchrdit.solver import create_solver
-
-        # Create a solver and add some layers
-        solver = create_solver()
-        solver.add_layer(material_name="Si", thickness=torch.tensor(0.5))
-
-        # Update the thickness of the first layer
-        solver.update_layer_thickness(layer_index=0, thickness=torch.tensor(0.75))
-
-        # For optimization, use a thickness parameter with gradients enabled
-        thickness_param = torch.tensor(0.5, requires_grad=True)
-        solver.update_layer_thickness(layer_index=0, thickness=thickness_param)
-        ```
-        """
-        self.layer_manager.update_layer_thickness(layer_index=layer_index, thickness=thickness)
-        # self.layer_manager.layers[layer_index].is_solved = False
-
-    def _get_bg(self, layer_index: int, param="er") -> torch.Tensor:
-        """_get_bg.
-
-        Returns the background matrix of the specified layer.
-
-        Args:
-            layer_index (int): layer_index
-            param:
-
-        Returns:
-            torch.Tensor:
-        """
-
-        ret_mat = None
-
-        if layer_index < self.layer_manager.nlayer:
-            if self.layer_manager.layers[layer_index].is_dispersive is True:
-                material_name = self.layer_manager.layers[layer_index].material_name
-                if param == "er":
-                    # ret_mat = torch.tensor(self._matlib[material_name].er).unsqueeze(1)\
-                    # .unsqueeze(1).repeat(1, self.grids[0], self.grids[1]).to(self.device).to(self.tcomplex)
-                    ret_mat = (
-                        self._matlib[material_name]
-                        .er.detach()
-                        .clone()
-                        .unsqueeze(1)
-                        .unsqueeze(1)
-                        .repeat(1, self.grids[0], self.grids[1])
-                        .to(self.device)
-                        .to(self.tcomplex)
-                    )
-                elif param == "ur":
-                    param_val = self._matlib[material_name].ur.detach().clone()
-                    ret_mat = param_val * torch.ones(
-                        size=(self.grids[0], self.grids[1]), dtype=self.tcomplex, device=self.device
-                    )
-                else:
-                    raise ValueError(f"Input parameter [{param}] is illeagal.")
-
-            else:
-                material_name = self.layer_manager.layers[layer_index].material_name
-                if param == "er":
-                    param_val = self._matlib[material_name].er.detach().clone()
-                elif param == "ur":
-                    param_val = self._matlib[material_name].ur.detach().clone()
-                else:
-                    raise ValueError(f"Input parameter [{param}] is illeagal.")
-
-                ret_mat = param_val * torch.ones(
-                    size=(self.grids[0], self.grids[1]), dtype=self.tcomplex, device=self.device
-                )
-        else:
-            raise ValueError("The index exceeds the max layer number.")
-
-        return ret_mat
-
-    def expand_dims(self, mat: torch.Tensor) -> Optional[torch.Tensor]:
-        """Function that expands the input matrix to a standard output dimension without layer information:
-            (n_freqs, n_harmonics_squared, n_harmonics_squared)
-
-        Args:
-            mat (torch.Tensor): input tensor matrxi
-        """
-
-        ret = None
-
-        if mat.ndim == 1 and mat.shape[0] == self.n_freqs:
-            # The input tensor with dimension (n_freqs)
-            ret = mat[:, None, None]
-        elif mat.ndim == 2:
-            # The input matrix with dimension (n_harmonics_squared, n_harmonics_squared)
-            ret = mat.unsqueeze(0).repeat(self.n_freqs, 1, 1)
-        else:
-            raise RuntimeError("Not Listed in the Case")
-
-        return ret
-
-    def _connect_external_regions(self, smat_global, matrices):
-        """Connect the global scattering matrix to reflection and transmission regions.
-
-        Args:
-            smat_global: The global scattering matrix to connect to external regions
-            matrices: Dictionary of matrices from setup_common_matrices
-
-        Returns:
-            tuple: (Updated global scattering matrix, V matrix for reflection, V matrix for transmission)
-        """
-
-        # Connect to reflection region — differentiable kz protection (see numerics.py)
-        inv_mat_lam_ref = safe_kz_reciprocal(matrices["mat_kz_ref"], dtype=self.tcomplex)
-
-        if self.layer_manager.is_ref_dispersive is False:
-            mat_v_ref = (1 / self.ur1) * blockmat2x2(
-                [
-                    [
-                        to_diag_util(matrices["mat_kx_ky"] * inv_mat_lam_ref, self.harmonics),
-                        to_diag_util((self.ur1 * self.er1 - matrices["mat_kx_kx"]) * inv_mat_lam_ref, self.harmonics),
-                    ],
-                    [
-                        to_diag_util((matrices["mat_ky_ky"] - self.ur1 * self.er1) * inv_mat_lam_ref, self.harmonics),
-                        -to_diag_util(matrices["mat_ky_kx"] * inv_mat_lam_ref, self.harmonics),
-                    ],
-                ]
-            )
-        else:
-            mat_v_ref = (1 / self.ur1) * blockmat2x2(
-                [
-                    [
-                        to_diag_util(matrices["mat_kx_ky"] * inv_mat_lam_ref, self.harmonics),
-                        to_diag_util(
-                            ((self.ur1 * self.er1)[:, None] - matrices["mat_kx_kx"]) * inv_mat_lam_ref, self.harmonics
-                        ),
-                    ],
-                    [
-                        to_diag_util(
-                            (matrices["mat_ky_ky"] - (self.ur1 * self.er1)[:, None]) * inv_mat_lam_ref, self.harmonics
-                        ),
-                        -to_diag_util(matrices["mat_ky_kx"] * inv_mat_lam_ref, self.harmonics),
-                    ],
-                ]
-            )
-
-        mat_w_ref = self.ident_mat_k2.unsqueeze(0).expand(self.n_freqs, -1, -1)
-
-        # Calculate reflection region scattering matrix
-        atw1 = mat_w_ref
-        atv1 = tsolve(matrices["mat_v0"], mat_v_ref)
-
-        mat_a_1 = atw1 + atv1
-        mat_b_1 = atw1 - atv1
-        inv_mat_a_1 = tinv(mat_a_1)
-        inv_mat_a_1_mat_b_1 = inv_mat_a_1 @ mat_b_1
-        smat_ref = SMatrix(
-            S11=-inv_mat_a_1_mat_b_1,
-            S12=2 * inv_mat_a_1,
-            S21=0.5 * (mat_a_1 - mat_b_1 @ inv_mat_a_1_mat_b_1),
-            S22=mat_b_1 @ inv_mat_a_1,
-        )
-
-        smat_global = redhstar(smat_ref, smat_global)
-
-        # Connect to transmission region — differentiable kz protection (see numerics.py)
-        inv_mat_lam_trn = safe_kz_reciprocal(matrices["mat_kz_trn"], dtype=self.tcomplex)
-
-        if self.layer_manager.is_trn_dispersive is False:
-            mat_v_trn = (1 / self.ur2) * blockmat2x2(
-                [
-                    [
-                        to_diag_util(matrices["mat_kx_ky"] * inv_mat_lam_trn, self.harmonics),
-                        to_diag_util((self.ur2 * self.er2 - matrices["mat_kx_kx"]) * inv_mat_lam_trn, self.harmonics),
-                    ],
-                    [
-                        to_diag_util((matrices["mat_ky_ky"] - self.ur2 * self.er2) * inv_mat_lam_trn, self.harmonics),
-                        -to_diag_util(matrices["mat_ky_kx"] * inv_mat_lam_trn, self.harmonics),
-                    ],
-                ]
-            )
-        else:
-            mat_v_trn = (1 / self.ur2) * blockmat2x2(
-                [
-                    [
-                        to_diag_util(matrices["mat_kx_ky"] * inv_mat_lam_trn, self.harmonics),
-                        to_diag_util(
-                            ((self.ur2 * self.er2)[:, None] - matrices["mat_kx_kx"]) * inv_mat_lam_trn, self.harmonics
-                        ),
-                    ],
-                    [
-                        to_diag_util(
-                            (matrices["mat_ky_ky"] - (self.ur2 * self.er2)[:, None]) * inv_mat_lam_trn, self.harmonics
-                        ),
-                        -to_diag_util(matrices["mat_ky_kx"] * inv_mat_lam_trn, self.harmonics),
-                    ],
-                ]
-            )
-
-        mat_w_trn = mat_w_ref
-
-        # Calculate transmission region scattering matrix
-        atw2 = mat_w_trn
-        atv2 = tsolve(matrices["mat_v0"], mat_v_trn)
-
-        mat_a_2 = atw2 + atv2
-        mat_b_2 = atw2 - atv2
-        inv_mat_a_2 = tinv(mat_a_2)
-        inv_mat_a_2_mat_b_2 = inv_mat_a_2 @ mat_b_2
-        smat_trn = SMatrix(
-            S11=mat_b_2 @ inv_mat_a_2,
-            S12=0.5 * (mat_a_2 - mat_b_2 @ inv_mat_a_2_mat_b_2),
-            S21=2 * inv_mat_a_2,
-            S22=-inv_mat_a_2_mat_b_2,
-        )
-
-        smat_global = redhstar(smat_global, smat_trn)
-
-        return smat_global, mat_v_ref, mat_v_trn
-
-    def _calculate_polarization(self, sources=None, kinc=None):
-        """Unified polarization calculation for both single and batched sources.
-
-        This method automatically detects whether it's dealing with a single source
-        or multiple sources and handles both cases using tensorized operations.
-
-        Args:
-            sources: Optional list of source dictionaries for batched processing.
-                    If None, uses self.src for single source processing.
-            kinc: Optional kinc tensor to use instead of self.kinc. If None, uses self.kinc.
-
-        Returns:
-            Dictionary containing:
-            - ate: TE polarization vectors
-            - atm: TM polarization vectors
-            - pol_vec: Combined polarization vectors
-            - esrc: Electric field source vectors
-
-            Shapes:
-            - Single source: (n_freqs, 3) for vectors, (n_freqs, 2*n_harmonics_squared) for esrc
-            - Batched sources: (n_sources, n_freqs, 3) for vectors, (n_sources, n_freqs, 2*n_harmonics_squared) for esrc
-        """
-        # Use passed kinc parameter or fall back to self.kinc
-        if kinc is None:
-            kinc = self.kinc
-
-        # Detect input type and normalize
-        if sources is None:
-            # Single source mode - use self.src
-            is_single = True
-            sources_list = [self.src]
-        elif isinstance(sources, dict):
-            # Single source passed as dict
-            is_single = True
-            sources_list = [sources]
-        else:
-            # Batched sources
-            is_single = False
-            sources_list = sources
-
-        n_sources = len(sources_list)
-        n_harmonics_squared = self.harmonics[0] * self.harmonics[1]
-
-        # Use tensorized parameter collection while preserving autograd
-        def _to_float_tensor_preserve(x):
-            if isinstance(x, torch.Tensor):
-                # Ensure device/dtype match; preserves graph
-                return x.to(device=self.device, dtype=self.tfloat)
-            else:
-                return torch.tensor(x, device=self.device, dtype=self.tfloat)
-
-        def _to_complex_tensor_preserve(x):
-            if isinstance(x, torch.Tensor):
-                return x.to(device=self.device, dtype=self.tcomplex)
-            else:
-                return torch.tensor(x, device=self.device, dtype=self.tcomplex)
-
-        theta_batch = torch.stack([_to_float_tensor_preserve(src["theta"]) for src in sources_list], dim=0)
-        pte_batch = torch.stack([_to_complex_tensor_preserve(src["pte"]) for src in sources_list], dim=0)
-        ptm_batch = torch.stack([_to_complex_tensor_preserve(src["ptm"]) for src in sources_list], dim=0)
-
-        # Initialize polarization tensors
-        ate_batch = torch.zeros(n_sources, self.n_freqs, 3, dtype=self.tcomplex, device=self.device)
-        atm_batch = torch.zeros(n_sources, self.n_freqs, 3, dtype=self.tcomplex, device=self.device)
-
-        # Vectorized normal incidence detection
-        normal_mask = torch.abs(theta_batch) < 1e-3
-
-        # Handle normal incidence cases
-        ate_batch[normal_mask, :, 1] = 1.0  # Default: [0, 1, 0] for y-direction
-
-        # Handle oblique incidence cases
-        oblique_mask = ~normal_mask
-        if oblique_mask.any():
-            # Ensure kinc is 3D (n_sources, n_freqs, 3) for masked indexing
-            if kinc.dim() == 2:
-                kinc_expanded = kinc[None, :, :].expand(n_sources, -1, -1)
-            else:
-                kinc_expanded = kinc
-            kinc_oblique = kinc_expanded[oblique_mask]
-
-            norm_vec = torch.tensor([0.0, 0.0, 1.0], dtype=self.tcomplex, device=self.device)
-            norm_vec_expanded = norm_vec[None, None, :].expand(oblique_mask.sum(), self.n_freqs, -1)
-
-            ate_oblique = torch.cross(kinc_oblique, norm_vec_expanded, dim=2)
-            ate_oblique_norm = torch.norm(ate_oblique, dim=2, keepdim=True)
-            ate_oblique = ate_oblique / (ate_oblique_norm + 1e-10)
-
-            ate_batch[oblique_mask] = ate_oblique
-
-        # Ensure kinc is 3D (n_sources, n_freqs, 3) for atm cross product
-        if kinc.dim() == 2:
-            kinc_for_atm = kinc[None, :, :].expand(n_sources, -1, -1)
-        else:
-            kinc_for_atm = kinc.expand(n_sources, -1, -1)
-
-        atm_batch = torch.cross(ate_batch, kinc_for_atm, dim=2)
-        atm_norm = torch.norm(atm_batch, dim=2, keepdim=True)
-        atm_batch = atm_batch / (atm_norm + 1e-10)
-
-        # Create polarization vectors using broadcasting
-        pte_vec = pte_batch[:, None, None] * ate_batch
-        ptm_vec = ptm_batch[:, None, None] * atm_batch
-
-        pol_vec = pte_vec + ptm_vec
-        pol_vec_norm = torch.norm(pol_vec, dim=2, keepdim=True)
-        pol_vec = pol_vec / (pol_vec_norm + 1e-10)
-
-        # Create electric field source vectors
-        delta = torch.zeros(
-            size=(n_sources, self.n_freqs, n_harmonics_squared), dtype=self.tcomplex, device=self.device
-        )
-        delta[:, :, n_harmonics_squared // 2] = 1.0
-
-        esrc = torch.zeros(
-            size=(n_sources, self.n_freqs, 2 * n_harmonics_squared), dtype=self.tcomplex, device=self.device
-        )
-        esrc[:, :, :n_harmonics_squared] = pol_vec[:, :, 0].unsqueeze(-1) * delta
-        esrc[:, :, n_harmonics_squared:] = pol_vec[:, :, 1].unsqueeze(-1) * delta
-
-        # Remove source dimension for single source case
-        if is_single:
-            ate_batch = ate_batch.squeeze(0)
-            atm_batch = atm_batch.squeeze(0)
-            pol_vec = pol_vec.squeeze(0)
-            esrc = esrc.squeeze(0)
-
-        return {"ate": ate_batch, "atm": atm_batch, "pol_vec": pol_vec, "esrc": esrc}
-
-    def _calculate_fields_and_efficiencies(
-        self, smat_global, matrices, kx_0, ky_0, mat_v_ref=None, mat_v_trn=None, kinc=None, sources=None
-    ):
-        """Calculate fields and diffraction efficiencies — vectorized over sources.
-
-        All inputs are 4D ``(n_sources, n_freqs, ...)``. All outputs in the
-        returned dict carry the same leading ``(n_sources, ...)`` dimension.
-
-        Args:
-            smat_global: Global scattering matrix dict, values ``(B, F, 2M, 2M)``
-            matrices: Common-matrix dict, values ``(B, F, ...)``
-            kx_0, ky_0: Wave vectors ``(B, F, H0, H1)``
-            mat_v_ref: V matrix for reflection region ``(B, F, 2M, 2M)``
-            mat_v_trn: V matrix for transmission region ``(B, F, 2M, 2M)``
-            kinc: Incident wave vectors ``(B, F, 3)``. Falls back to ``self.kinc``.
-            sources: List of source dicts (for batched polarization). Falls back to ``self.src``.
-
-        Returns:
-            dict: Fields and efficiencies with leading ``(B, ...)`` dimension.
-        """
-        if kinc is None:
-            kinc = self.kinc
-        n_sources = kinc.shape[0]
-
-        n_harmonics_squared = self.harmonics[0] * self.harmonics[1]
-
-        # Batched polarization: esrc shape (B, F, 2*H²)
-        polarization_data = self._calculate_polarization(sources=sources, kinc=kinc)
-        esrc = polarization_data["esrc"]
-
-        # W_ref = identity, shape (B, F, 2M, 2M)
-        M2 = 2 * n_harmonics_squared
-        mat_w_ref = self.ident_mat_k2.reshape(1, 1, M2, M2).expand(n_sources, self.n_freqs, -1, -1)
-        mat_w_trn = mat_w_ref
-
-        csrc = tsolve(mat_w_ref, esrc.unsqueeze(-1))  # (B, F, 2M, 1)
-
-        # Reflected fields
-        cref = smat_global.S11 @ csrc  # (B, F, 2M, 1)
-        eref = mat_w_ref @ cref
-
-        ref_field_x = eref[:, :, 0:n_harmonics_squared, :]
-        ref_field_y = eref[:, :, n_harmonics_squared:M2, :]
-
-        min_kz = getattr(self, "min_kinc_z", 1e-3)
-
-        mat_kz_ref_protected = softplus_magnitude_floor(matrices["mat_kz_ref"], min_kz, beta=100)
-        kz_ref_diag = to_diag_util(mat_kz_ref_protected, self.harmonics)
-        ref_field_z = (
-            -matrices["mat_kx_diag"] @ tsolve(kz_ref_diag, ref_field_x)
-            - matrices["mat_ky_diag"] @ tsolve(kz_ref_diag, ref_field_y)
-        )
-
-        # Transmitted fields
-        ctrn = smat_global.S21 @ csrc
-        etrn = mat_w_trn @ ctrn
-        trn_field_x = etrn[:, :, 0:n_harmonics_squared, :]
-        trn_field_y = etrn[:, :, n_harmonics_squared:M2, :]
-
-        mat_kz_trn_protected = softplus_magnitude_floor(matrices["mat_kz_trn"], min_kz, beta=100)
-        kz_trn_diag = to_diag_util(mat_kz_trn_protected, self.harmonics)
-        trn_field_z = (
-            -matrices["mat_kx_diag"] @ tsolve(kz_trn_diag, trn_field_x)
-            - matrices["mat_ky_diag"] @ tsolve(kz_trn_diag, trn_field_y)
-        )
-
-        # Diffraction efficiencies — kinc shape (B, F, 3)
-        kinc_z = kinc[:, :, 2]  # (B, F)
-        min_kinc_z = getattr(self, "min_kinc_z", 1e-3)
-        kinc_z_protected = softplus_floor(torch.real(kinc_z), min_kinc_z, beta=100)  # (B, F)
-
-        def _field_intensity(fx, fy, fz):
-            return (
-                torch.real(fx) ** 2 + torch.imag(fx) ** 2
-                + torch.real(fy) ** 2 + torch.imag(fy) ** 2
-                + torch.real(fz) ** 2 + torch.imag(fz) ** 2
-            )
-
-        H0, H1 = self.harmonics
-        # kinc_z_protected[:, :, None, None] broadcasts to (B, F, H², 1)
-        ref_diff_efficiency = torch.reshape(
-            torch.real(
-                self.ur2 / self.ur1 * kz_ref_diag / kinc_z_protected[:, :, None, None]
-            ) @ _field_intensity(ref_field_x, ref_field_y, ref_field_z),
-            shape=(n_sources, self.n_freqs, H0, H1),
-        ).transpose(dim0=-2, dim1=-1)
-
-        trn_diff_efficiency = torch.reshape(
-            torch.real(
-                self.ur1 / self.ur2 * kz_trn_diag / kinc_z_protected[:, :, None, None]
-            ) @ _field_intensity(trn_field_x, trn_field_y, trn_field_z),
-            shape=(n_sources, self.n_freqs, H0, H1),
-        ).transpose(dim0=-2, dim1=-1)
-
-        total_ref_efficiency = torch.sum(ref_diff_efficiency, dim=(-1, -2))  # (B, F)
-        total_trn_efficiency = torch.sum(trn_diff_efficiency, dim=(-1, -2))
-
-        # Magnetic field coefficients
-        if mat_v_ref is not None and mat_v_trn is not None:
-            uref = -mat_v_ref @ cref
-            ref_mag_x = uref[:, :, 0:n_harmonics_squared, :]
-            ref_mag_y = uref[:, :, n_harmonics_squared:M2, :]
-            ref_mag_z = (
-                -matrices["mat_kx_diag"] @ tsolve(kz_ref_diag, ref_mag_x)
-                - matrices["mat_ky_diag"] @ tsolve(kz_ref_diag, ref_mag_y)
-            )
-
-            utrn = mat_v_trn @ ctrn
-            trn_mag_x = utrn[:, :, 0:n_harmonics_squared, :]
-            trn_mag_y = utrn[:, :, n_harmonics_squared:M2, :]
-            trn_mag_z = (
-                -matrices["mat_kx_diag"] @ tsolve(kz_trn_diag, trn_mag_x)
-                - matrices["mat_ky_diag"] @ tsolve(kz_trn_diag, trn_mag_y)
-            )
-        else:
-            ref_mag_x = ref_mag_y = ref_mag_z = None
-            trn_mag_x = trn_mag_y = trn_mag_z = None
-
-        return {
-            "ref_s_x": ref_field_x, "ref_s_y": ref_field_y, "ref_s_z": ref_field_z,
-            "trn_s_x": trn_field_x, "trn_s_y": trn_field_y, "trn_s_z": trn_field_z,
-            "ref_u_x": ref_mag_x, "ref_u_y": ref_mag_y, "ref_u_z": ref_mag_z,
-            "trn_u_x": trn_mag_x, "trn_u_y": trn_mag_y, "trn_u_z": trn_mag_z,
-            "ref_diff_efficiency": ref_diff_efficiency,
-            "trn_diff_efficiency": trn_diff_efficiency,
-            "total_ref_efficiency": total_ref_efficiency,
-            "total_trn_efficiency": total_trn_efficiency,
-        }
-
 
 class RCWASolver(FourierBaseSolver):
     """Rigorous Coupled-Wave Analysis (RCWA) Solver.
